Isolate the File Component and create the views and urls to get the cells of a specific file
