+++ personalwiki/js/Cell.tsx
@@ -44,23 +44,26 @@ class Cell extends React.Component {
 
 			// INPUT MODE
 			return(
-				<div className={ `cell cell-input ${(this.state.data==='') ? 'cell-empty' : '' }` }
-             contentEditable
-						 tabIndex={0}
-             data-id={this.props.id}
-						 data-uuid={this.props.uuid}
-						 ref={this.cell_text}
-						 onBlur={ () => {
-             	this.setState({selected: false, editing: false, data: this.cell_text.current.innerText.replaceAll('\n', '\n\n')});
-							if(this.cell_text.current.innerText.replaceAll('\n', '\n\n') !== ''){
-								this.props.alert_action( (this.b_initial_post) ? Core.cell_data_methods.POST
-							 	                                               : Core.cell_data_methods.PATCH, 
-																					this.props.id, this.cell_text.current.innerText.replaceAll('\n', '\n\n'));
-								this.cell_text.current.innerText = '';
-								if(this.b_initial_post) this.b_initial_post = false;
-							}
-						 }}>
-					{this.state.data.replaceAll('\n\n', '\r\n')}
+				<div className='cell-wrapper'>
+					<div className={ `cell cell-input ${(this.state.data==='') ? 'cell-empty' : '' }` }
+         	    contentEditable
+							 tabIndex={0}
+         	    data-id={this.props.id}
+							 data-uuid={this.props.uuid}
+							 ref={this.cell_text}
+							 onBlur={ () => {
+         	    	this.setState({selected: false, editing: false, data: this.cell_text.current.innerText.replaceAll('\n', '\n\n')});
+								if(this.cell_text.current.innerText.replaceAll('\n', '\n\n') !== ''){
+									this.props.alert_action( (this.b_initial_post) ? Core.cell_data_methods.POST
+								 	                                               : Core.cell_data_methods.PATCH, 
+																						this.props.id, this.cell_text.current.innerText.replaceAll('\n', '\n\n'));
+									this.cell_text.current.innerText = '';
+									if(this.b_initial_post) this.b_initial_post = false;
+								}
+							 }}>
+						{this.state.data.replaceAll('\n\n', '\r\n')}
+					</div>
+					<p className='cell-id-prompt cell-id-prompt__input'> &#9664; [{ this.props.id + 1}]</p>
 				</div>
 			)
 		} else {
@@ -79,13 +82,14 @@ class Cell extends React.Component {
 					</div>
 					<div className='cell-selected-options'>
 						<button type='button' onClick={ () => this.alert_parent( Core.cell_ui_methods.CELL_MOVE_UP, this.props.id)}>
-							<img src={IMG_ARROW} alt='move cell up' /></button>
+							<p>&#9650;</p></button>
 						<button type='button' onClick={ () => this.alert_parent( Core.cell_data_methods.DELETE, this.props.id)}>
+							{ //<p>&#10006;</p></button> }
 							<img src={IMG_TRASH} alt='delete cell' /></button>
 						<button type='button' onClick={ () => this.alert_parent( Core.cell_ui_methods.CELL_MOVE_DOWN, this.props.id)}>
-							<img src={IMG_ARROW} alt='move cell down' /></button>
+							<p>&#9660;</p></button>
 					</div>
-					<p className='cell-id-prompt'> &#9664; [{ this.props.id + 1}]</p>
+					<p className={ this.state.selected ? 'cell-id-prompt cell-id-prompt__selected' : 'cell-id-prompt'}> &#9664; [{ this.props.id + 1}]</p>
 				</div>
 			)
 		}
