+++ personalwiki/templates/dirs.html
@@ -1,0 +1,23 @@
+
+
+
+{% load static %}
+
+<!doctype html>
+<html>
+	<head>
+		<meta charset='UTF-8' />
+		<meta name='viewport' content='width=device-width, initial-scale=1.0' />
+		<meta name='description' content='personalwiki 0.1' />
+		<title>Directory</title>
+		<link rel='stylesheet' href="{% static './styles/index.css' %}" />
+		<link rel='stylesheet' href="{% static './styles/markdown.css' %}" />
+	</head>
+
+	<body>
+		<div id="root"></div>
+
+		<script src="https://unpkg.com/react@18/umd/react.development.js" crossorigin></script>
+    <script src="https://unpkg.com/react-dom@18/umd/react-dom.development.js" crossorigin></script>
+	</body>
+</html>
