+++ personalwiki/wikiapp/models.py
@@ -34,7 +34,7 @@ class Cell(models.Model):
     data = models.TextField()
     uuid = models.UUIDField(primary_key=True, unique=True, default=uuid.uuid4, editable=False)
     uhash = models.CharField(max_length=120, default="NULL")
-    #main_file = models.ForeignKey(File, on_delete=models.SET_NULL, null=True, blank=True)
+    main_file = models.ForeignKey(File, on_delete=models.SET_NULL, null=True, blank=True)
 
     def __str__(self):
         return self.data
