+++ personalwiki/wikiapp/static/styles/markdown.css
@@ -26,7 +26,11 @@
 
 .cell code {
 	background-color: #eee;
-	padding: 2px 5px;
+	padding: 2px 6px;
 	border-radius: 5px;
+}
+
+.cell pre code {
+	padding: 0px;
 }
